Update setup-controller.sh
Use /dev/shm rather than /tmp/setup for storing the vmdk file and its compressed version until they are copied over with the "glance" command, which stores the vmdk file under /var/lib/glance. The reason for this is I encountered an out of disk space error when glance tried to copy the image to the /var/lib/glance folder. Since there is far more memory on the controller node than there is for the root partition on the controller node, using /dev/shm should work and not throw an error.